apps/learnetic_monitoreo_mcourserpro/priv/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cachetools==2.0.0
cffi==1.10.0
cryptography==1.8.1
enum34==1.1.6
erlang-py==1.7.0
et-xmlfile==1.0.1
gdata==2.0.18
google-api-python-client==1.6.2
google-auth==0.10.0
google-auth-httplib2==0.0.2
google-cloud-bigquery==0.24.0
google-cloud-core==0.24.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
jdcal==1.3
numpy==1.12.1
oauth2client==4.0.0
packaging==16.8
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyMySQL==0.7.11
pyOpenSSL==16.2.0
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
scipy==0.19.0
six==1.10.0
uritemplate==3.0.0
